+++ backend/src/repositories/chat.py
@@ -4,26 +4,22 @@ from sqlalchemy import select
 from sqlalchemy.orm import joinedload
 
 from src.models import Chat as ChatBody
-from src.models.sql.chat import Chat, Image, Message
+from src.models.sql.chat import Chat, Image, Message, Source
 
 class ChatRepository:
     def __init__(self, request, db):
         self.db = db
         self.user_id = request.state.user_id
-        self.soft_delete = True
     
     async def list(self):
         stmt = (
             select(Chat)
-            .options(joinedload(Chat.messages).load_only(Message.id, Message.role, Message.content, Message.created_at))
+            .options(joinedload(Chat.messages).load_only(Message.id, Message.role, Message.content, Message.model, Message.created_at))
             .where(Chat.user_id == self.user_id)
             .order_by(Chat.updated_at.desc())
         )
 
-        if self.soft_delete:
-            stmt = stmt.where(Chat.deleted_at.is_(None))
-
-        print(str(stmt.compile(compile_kwargs={"literal_binds": True})))
+        # print(str(stmt.compile(compile_kwargs={"literal_binds": True})))
         result = await self.db.execute(stmt)
         chats = result.unique().scalars().all()
 
@@ -31,7 +27,7 @@ class ChatRepository:
             {
                 "id": chat.id,
                 "messages": [
-                    {"role": message.role, "content": message.content, "created_at": message.created_at.isoformat()}
+                    {"role": message.role, "content": message.content, "model": message.model, "created_at": message.created_at.isoformat()}
                     for message in sorted(chat.messages, key=lambda m: m.created_at)  # Sort messages here if not sorted in model
                 ],
                 "created_at": chat.created_at.isoformat(),
@@ -52,17 +48,32 @@ class ChatRepository:
                 await self.db.flush()
 
                 for message in chat.messages:
-                    images = message.get("images", [])
+                    
                     new_message = Message(chat_id=new_chat.id, 
                                           role=message["role"], 
-                                          content=message["content"], 
+                                          content=message["content"],
+                                          model=message.get("model"), 
                                           created_at=datetime.utcnow())
                     self.db.add(new_message)
                     await self.db.flush()
+                    
+                    images = message.get("images", [])
                     for image in images:
-                        new_image = Image(message_id=new_message.id, 
-                                        content=image)
+                        new_image = Image(
+                            message_id=new_message.id, 
+                            content=image
+                        )
                         self.db.add(new_image)
+                        
+                    sources = message.get("sources", [])
+                    for source in sources:
+                        new_source = Source(
+                            message_id=new_message.id,
+                            name=source.get("name"), 
+                            type=source.get("type"),
+                            src=source.get("src"),
+                        )
+                        self.db.add(new_source)
                     
                     
                 return {"id": new_chat.id}
@@ -75,7 +86,8 @@ class ChatRepository:
         stmt = (
             select(Chat)
             .options(
-                joinedload(Chat.messages).joinedload(Message.images)  # Load images for each message
+                joinedload(Chat.messages).joinedload(Message.images),
+                joinedload(Chat.messages).joinedload(Message.sources)
             )
             .where(Chat.id == chat_id, Chat.user_id == self.user_id)
         )
@@ -88,9 +100,19 @@ class ChatRepository:
                     {
                         "role": message.role, 
                         "content": message.content,
+                        "model": message.model or None,
                         "images": [
                             image.content for image in message.images
-                        ]
+                        ],
+                        "sources": [
+                            {
+                                "name": source.name,
+                                "type": source.type,
+                                "src": source.src
+                            }
+                            for source in message.sources
+                        ],
+                        "created_at": message.created_at.isoformat()
                     }
                     for message in chat.messages
                 ],
@@ -101,61 +123,63 @@ class ChatRepository:
         return None
     
     async def update(self, chat_id: int, updates: ChatBody):
-        try:
-            async with self.db.begin():
+        async with self.db.begin() as transaction:
+            try:
                 # Fetch the chat and its messages
-                stmt = select(Chat).options(joinedload(Chat.messages)).where(Chat.id == chat_id, 
-                                                                             Chat.user_id == self.user_id)
+                stmt = select(Chat).options(joinedload(Chat.messages)).where(Chat.id == chat_id, Chat.user_id == self.user_id)
                 result = await self.db.execute(stmt)
                 chat = result.scalars().first()
 
                 if chat:
-                    
-                    ## This makes it work like mongo db
-                    ## was added as type of placeholder. Simplifies 
-                    ## messages to work more like a document
-                    if True: 
-                        # Delete all current messages
-                        for message in chat.messages:
-                            await self.db.delete(message)
+                    # Soft delete all current messages
+                    for message in chat.messages:
+                        await self.db.delete(message)
 
                     # Set the updated_at field
                     chat.updated_at = datetime.utcnow()
 
-                    messages = updates.messages
                     # Add new messages
+                    messages = updates.messages
                     if messages:
-                         for message in messages:
+                        for message_data in messages:
                             new_message = Message(chat_id=chat_id, 
-                                                role=message["role"], 
-                                                content=message["content"], 
-                                                created_at=datetime.utcnow())
+                                                  role=message_data["role"], 
+                                                  content=message_data["content"], 
+                                                  model=message_data.get("model"), 
+                                                  created_at=datetime.utcnow())
                             self.db.add(new_message)
+                            await self.db.flush()
+
+                            images = message_data.get("images", [])
+                            for image in images:
+                                new_image = Image(
+                                    message_id=new_message.id, 
+                                    content=image
+                                )
+                                self.db.add(new_image)
+
+                            sources = message_data.get("sources", [])
+                            for source in sources:
+                                new_source = Source(
+                                    message_id=new_message.id,
+                                    name=source.get("name"), 
+                                    type=source.get("type"),
+                                    src=source.get("src"),
+                                )
+                                self.db.add(new_source)
 
                     # Return the updated chat information
                     return {"id": chat.id, "updated_at": chat.updated_at.isoformat()}
-
-        except Exception as e:
-            await self.db.rollback()  # Rollback in case of exception
-            raise
-        finally:
-            if self.db.is_active:
-                await self.db.rollback()
+            except Exception as e:
+                await transaction.rollback()
+                raise e
 
     async def delete(self, chat_id: int):
         async with self.db.begin():
-            stmt = select(Chat).options(joinedload(Chat.messages)).where(Chat.id == chat_id, 
-                                                                         Chat.user_id == self.user_id)
+            stmt = select(Chat).where(Chat.id == chat_id, Chat.user_id == self.user_id)
             result = await self.db.execute(stmt)
             chat = result.scalars().first()
             if chat:
-                if self.soft_delete:
-                    ## Soft delete of chat and associated messages
-                    chat.deleted_at = datetime.utcnow()
-                    for message in chat.messages:
-                        message.deleted_at = datetime.utcnow()
-                else:
-                    ## Full delete of chat and associated messages
-                    await self.db.delete(chat)
+                await self.db.delete(chat)
                 return True
             return None