+++ Changelog.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## v0.0.1
+### Fixed
+  - bugfix/33-groq-rag-not-working (2024-05-19)
 ### Changed 
+  - feature/35-db-indexes (2024-05-19)
+  - feature/23-message-include-model-used (2024-05-19)
+  - feature/29-storing-sources-on-messages (2024-05-19)
   - feature/27-files-to-be-shown-in-thread (2024-05-18)
   - feature/1-add-storage-and-db-services
   - development